HyperliquidExchange/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
httpx
hyperliquid-python-sdk
eth-account